src/detection/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
onnxruntime>=1.10.0
google-cloud-pubsub>=2.13.0
lap>=0.4.0
cython-bbox>=0.1.3